requirements.txt
python-telegram-bot[job-queue]==21.6
groq==0.11.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
